Rename « S3 » service to « object_store ».
