+++ src/backend/main/go.backends/store/cassandra/cassandra.go
@@ -5,7 +5,7 @@
 package store
 
 import (
-	"github.com/CaliOpen/Caliopen/src/backend/main/go.backends/store/S3"
+	"github.com/CaliOpen/Caliopen/src/backend/main/go.backends/store/object_store"
 	log "github.com/Sirupsen/logrus"
 	"github.com/gocassa/gocassa"
 	"github.com/gocql/gocql"
@@ -17,7 +17,7 @@ type (
 		CassandraConfig
 		Session   *gocql.Session
 		IKeyspace gocassa.KeySpace //gocassa keyspace interface
-		S3Backend S3.S3Backend
+		S3Backend object_store.S3Backend
 	}
 
 	CassandraConfig struct {
@@ -26,7 +26,7 @@ type (
 		Consistency gocql.Consistency `mapstructure:"consistency_level"`
 		SizeLimit   uint64            `mapstructure:"raw_size_limit"` // max size to store (in bytes)
 		WithS3      bool              // whether to use a S3 compatible storage engine for objects above SizeLimit
-		S3.S3Config
+		object_store.S3Config
 	}
 )
 
@@ -62,7 +62,7 @@ func (cb *CassandraBackend) initialize(config CassandraConfig) (err error) {
 	cb.IKeyspace = connection.KeySpace(cb.Keyspace)
 
 	if config.WithS3 {
-		cb.S3Backend, err = S3.InitializeS3Backend(config.S3Config)
+		cb.S3Backend, err = object_store.InitializeS3Backend(config.S3Config)
 		if err != nil {
 			return err
 		}
