+++ src/backend/main/go.backends/store/object_store/emails.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a GNU AFFERO GENERAL PUBLIC
 // license (AGPL) that can be found in the LICENSE file.
 
-package S3
+package object_store
 
 import (
 	"fmt"
