+++ src/backend/main/go.backends/store/object_store/minio.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a GNU AFFERO GENERAL PUBLIC
 // license (AGPL) that can be found in the LICENSE file.
 
-package S3
+package object_store
 
 import (
 	obj "github.com/CaliOpen/Caliopen/src/backend/defs/go-objects"
